backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
click==8.1.7
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-otp==1.5.4
django-stubs==5.1.1
django-stubs-ext==5.1.1
django-types==0.19.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.9
PyJWT==2.9.0
pyotp==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tomli==2.0.2
types-psycopg2==2.9.21.20241019
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
uritemplate==4.1.1
django-oauth-toolkit==3.0.1
